competitive-intel/requirements.txt
requests>=2.31,<3
beautifulsoup4>=4.12,<5
feedparser>=6.0,<7
PyYAML>=6.0,<7
python-dotenv>=1.0,<2
